Created observation web scraper
This web scraper function downloads all observations for the current day from waarneming.nl. The observations are continuously compared to a life list set by the user id parameter.
This is to reduce running time of the function.
The function skips all observations of domesticated, hybrid and sub- species as well as skipping uncertain observations or ones under GPS-embargo.

The relevant information is stored per observation and subsequently appended to a dataframe stored as RDS file. This file gets called by app.R